Upgrade pip version in Makefile
The pip installed in virtaulenv is using the source tarfile from
virtaulenv package, but we uses apt to install virtaulenv, which is the
not-updated version, and that causes we have pip 9.0.1 in our venv.

Old pip version may cause some python dependencies build fail, and the
updated pip version in virtaulenv can avoid this problem.

Change-Id: I1209754d48ec7f84623d8a21bb9003cce6f0924f